Move boolean flag to const Use hook for claim function
