+++ pages/badge/preview/[badgeId].tsx
@@ -8,9 +8,9 @@ import { useTranslation } from 'next-export-i18n'
 import SafeSuspense, { withPageGenericSuspense } from '@/src/components/helpers/SafeSuspense'
 import useBadgeIdParam from '@/src/hooks/nextjs/useBadgeIdParam'
 import useBadgeById from '@/src/hooks/subgraph/useBadgeById'
+import useBadgeClaim from '@/src/hooks/theBadge/useBadgeClaim'
 import useBadgeHelpers, { ReviewBadge } from '@/src/hooks/theBadge/useBadgeHelpers'
-import { useContractInstance } from '@/src/hooks/useContractInstance'
-import useTransaction from '@/src/hooks/useTransaction'
+import { useSizeSM } from '@/src/hooks/useSize'
 import BadgeOwnedPreview from '@/src/pagePartials/badge/preview/BadgeOwnedPreview'
 import BadgeOwnerPreview from '@/src/pagePartials/badge/preview/BadgeOwnerPreview'
 import BadgeStatusAndEvidence from '@/src/pagePartials/badge/preview/BadgeStatusAndEvidence'
@@ -20,18 +20,16 @@ import { useColorMode } from '@/src/providers/themeProvider'
 import { useWeb3Connection } from '@/src/providers/web3ConnectionProvider'
 import { generateMintUrl } from '@/src/utils/navigation/generateUrl'
 import { BadgeStatus } from '@/types/generated/subgraph'
-import { TheBadge__factory } from '@/types/generated/typechain'
 import { NextPageWithLayout } from '@/types/next'
 
 const ViewBadge: NextPageWithLayout = () => {
   const { t } = useTranslation()
   const { address } = useWeb3Connection()
-  const { sendTx } = useTransaction()
   const { curate } = useCurateProvider()
   const { getBadgeReviewStatus } = useBadgeHelpers()
   const router = useRouter()
   const { mode } = useColorMode()
-  const theBadge = useContractInstance(TheBadge__factory, 'TheBadge')
+  const handleClaimBadge = useBadgeClaim()
   const isMobile = useSizeSM()
 
   const badgeId = useBadgeIdParam()
@@ -52,12 +50,15 @@ const ViewBadge: NextPageWithLayout = () => {
     badge as ReviewBadge,
   )
 
-  async function handleClaimBadge(badgeId: string) {
-    const transaction = await sendTx(() => theBadge.claim(badgeId, '0x'))
-    if (transaction) {
-      await transaction.wait()
-    }
-  }
+  // Show mint button if this is not the own badge
+  const showMintButton = address !== ownerAddress
+
+  // Show curate button if this is not the own badge and its not already challenged */
+  const showCurateButton = address !== ownerAddress && badgeStatus !== BadgeStatus.Challenged
+
+  // Show claim button if it is an own badge, it has status requested and the review time finished
+  const showClaimButton =
+    address === ownerAddress && badgeStatus === BadgeStatus.Requested && badgeReviewTimeFinished
 
   return (
     <Box sx={{ position: 'relative' }}>
@@ -73,79 +74,74 @@ const ViewBadge: NextPageWithLayout = () => {
               justifyContent="space-between"
               maxWidth={300}
             >
-              {/* Show mint button if this is not the own badge */}
-            {address !== ownerAddress ? (
-              <ButtonV2
-                backgroundColor={colors.transparent}
-                disabled={address === ownerAddress}
-                fontColor={mode === 'light' ? colors.blackText : colors.white}
-                onClick={() =>
-                  router.push(generateMintUrl(badge?.badgeModel.controllerType, badgeModelId))
-                }
-                sx={{
-                  borderRadius: '10px',
-                  fontSize: '11px !important',
-                  padding: '0.5rem 1rem !important',
-                  height: 'fit-content !important',
-                  lineHeight: '14px',
-                  fontWeight: 700,
-                  boxShadow: 'none',
-                  textTransform: 'uppercase',
-                }}
-              >
-                {t('badge.mintButton')}
-              </ButtonV2>
-            ) : null}
+              {showMintButton && (
+                <ButtonV2
+                  backgroundColor={colors.transparent}
+                  disabled={address === ownerAddress}
+                  fontColor={mode === 'light' ? colors.blackText : colors.white}
+                  onClick={() =>
+                    router.push(generateMintUrl(badge?.badgeModel.controllerType, badgeModelId))
+                  }
+                  sx={{
+                    borderRadius: '10px',
+                    fontSize: '11px !important',
+                    padding: '0.5rem 1rem !important',
+                    height: 'fit-content !important',
+                    lineHeight: '14px',
+                    fontWeight: 700,
+                    boxShadow: 'none',
+                    textTransform: 'uppercase',
+                  }}
+                >
+                  {t('badge.mintButton')}
+                </ButtonV2>
+              )}
 
-              {/* Show curate button if this is not the own badge and its not already challenged */}
-            {address !== ownerAddress && badgeStatus !== BadgeStatus.Challenged ? (
-              <ButtonV2
-                backgroundColor={colors.greenLogo}
-                disabled={address === ownerAddress}
-                fontColor={colors.blackText}
-                onClick={() => curate(badgeId)}
-                sx={{
-                  borderRadius: '10px',
-                  fontSize: '11px !important',
-                  padding: '0.5rem 1rem !important',
-                  height: 'fit-content !important',
-                  lineHeight: '14px',
-                  fontWeight: 700,
-                  boxShadow: 'none',
-                  textTransform: 'uppercase',
-                }}
-                variant="contained"
-              >
-                {t('badge.curateButton')}
-              </ButtonV2>
-            ) : null}
+              {showCurateButton && (
+                <ButtonV2
+                  backgroundColor={colors.greenLogo}
+                  disabled={address === ownerAddress}
+                  fontColor={colors.blackText}
+                  onClick={() => curate(badgeId)}
+                  sx={{
+                    borderRadius: '10px',
+                    fontSize: '11px !important',
+                    padding: '0.5rem 1rem !important',
+                    height: 'fit-content !important',
+                    lineHeight: '14px',
+                    fontWeight: 700,
+                    boxShadow: 'none',
+                    textTransform: 'uppercase',
+                  }}
+                  variant="contained"
+                >
+                  {t('badge.curateButton')}
+                </ButtonV2>
+              )}
 
-            {/* Show claim button if it is an own badge, it has status requested and the review time finished */}
-            {address === ownerAddress &&
-            badgeStatus === BadgeStatus.Requested &&
-            badgeReviewTimeFinished ? (
-              <ButtonV2
-                backgroundColor={colors.blue}
-                disabled={!badgeReviewTimeFinished}
-                fontColor={colors.white}
-                onClick={() => handleClaimBadge(badge.id)}
-                sx={{
-                  width: '100%',
-                  height: 'fit-content !important',
-                  marginTop: '1rem',
-                  padding: '0.5rem 1rem !important',
-                  borderRadius: '10px',
-                  fontSize: '15px !important',
-                  lineHeight: '15px',
-                  fontWeight: 700,
-                  boxShadow: 'none',
-                  textTransform: 'uppercase',
-                }}
-                variant="contained"
-              >
-                {t('badge.claimButton')}
-              </ButtonV2>
-            ) : null}
+              {showClaimButton && (
+                <ButtonV2
+                  backgroundColor={colors.blue}
+                  disabled={!badgeReviewTimeFinished}
+                  fontColor={colors.white}
+                  onClick={() => handleClaimBadge(badge.id)}
+                  sx={{
+                    width: '100%',
+                    height: 'fit-content !important',
+                    marginTop: '1rem',
+                    padding: '0.5rem 1rem !important',
+                    borderRadius: '10px',
+                    fontSize: '15px !important',
+                    lineHeight: '15px',
+                    fontWeight: 700,
+                    boxShadow: 'none',
+                    textTransform: 'uppercase',
+                  }}
+                  variant="contained"
+                >
+                  {t('badge.claimButton')}
+                </ButtonV2>
+              )}
             </Box>
           )}
           <SafeSuspense>
